+++ packages/clerk-js/src/ui/elements/IdentityPreview.tsx
@@ -60,7 +60,7 @@ const IdentifierContainer = (props: React.PropsWithChildren) => {
   return (
     <Text
       elementDescriptor={descriptors.identityPreviewText}
-      colorScheme='neutral'
+      colorScheme='secondary'
       truncate
       {...props}
     />