+++ packages/types/src/localization.ts
@@ -47,6 +47,10 @@ type _LocalizationResource = {
   formFieldLabel__organizationDomainDeletePending: LocalizationValue;
   formFieldLabel__confirmDeletion: LocalizationValue;
   formFieldLabel__role: LocalizationValue;
+  /**
+   * @experimental
+   */
+  __experimental_formFieldLabel__passkeyName: LocalizationValue;
   formFieldInputPlaceholder__emailAddress: LocalizationValue;
   formFieldInputPlaceholder__emailAddresses: LocalizationValue;
   formFieldInputPlaceholder__phoneNumber: LocalizationValue;
@@ -319,6 +323,14 @@ type _LocalizationResource = {
         primaryButton__updatePassword: LocalizationValue;
         primaryButton__setPassword: LocalizationValue;
       };
+      /**
+       * @experimental
+       */
+      __experimental_passkeysSection: {
+        title: LocalizationValue;
+        menuAction__rename: LocalizationValue;
+        menuAction__destructive: LocalizationValue;
+      };
       mfaSection: {
         title: LocalizationValue;
         primaryButton: LocalizationValue;
@@ -387,6 +399,17 @@ type _LocalizationResource = {
         messageLine1: LocalizationValue;
         messageLine2: LocalizationValue;
         successMessage: LocalizationValue;
+      };
+    };
+    /**
+     * @experimental
+     */
+    __experimental_passkeyScreen: {
+      title__rename: LocalizationValue;
+      subtitle__rename: LocalizationValue;
+      removeResource: {
+        title: LocalizationValue;
+        messageLine1: LocalizationValue;
       };
     };
     phoneNumberPage: {