+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Change Log
 
+## 1.0.0-beta.34
+
+### Minor Changes
+
+- Re-export everything from `@clerk/backend` in `@clerk/fastify` to support common backend types and functionality without adding `@clerk/backend` as dependency. ([#2968](https://github.com/clerk/javascript/pull/2968)) by [@dimkl](https://github.com/dimkl)
+
+  New exports:
+
+  - `verifyToken()`
+
+  New exported types:
+
+  - `ClerkOptions`
+  - `ClerkClient`
+  - `OrganizationMembershipRole`
+  - `VerifyTokenOptions`
+  - `WebhookEvent`
+  - `WebhookEventType`
+  - `AllowlistIdentifier`
+  - `Client`
+  - `EmailAddress`
+  - `ExternalAccount`
+  - `Invitation`
+  - `OauthAccessToken`
+  - `Organization`
+  - `OrganizationInvitation`
+  - `OrganizationMembership`
+  - `OrganizationMembershipPublicUserData`
+  - `PhoneNumber`
+  - `Session`
+  - `SignInToken`
+  - `SMSMessage`
+  - `Token`
+  - `User`
+
+- Expose debug headers in response for handshake / signed-out states from SDKs using headers returned from `authenticateRequest()` ([#2898](https://github.com/clerk/javascript/pull/2898)) by [@dimkl](https://github.com/dimkl)
+
+### Patch Changes
+
+- Updated dependencies [[`2964f8a47`](https://github.com/clerk/javascript/commit/2964f8a47e473fa8457a27104adb4d008613a0e3)]:
+  - @clerk/backend@1.0.0-beta.28
+
 ## 1.0.0-beta.33
 
 ### Patch Changes